lookup4 validation in AVX-512
 So we now have utf8 to utf16 working (both validating and
non-validating) + the lookup4 algorithm ported to AVX-512.

We pause after each AVX-512 call to make sure frequency throttling does
not affect other calls.
