+++ benchmarks/src/benchmark_base.cpp
@@ -96,7 +96,7 @@ namespace simdutf::benchmarks {
     void BenchmarkBase::print_summary(const event_aggregate& all, double data_size, double character_count) const {
         const double gbs = data_size / all.best.elapsed_ns();
         const double gcs = character_count / all.best.elapsed_ns();
-        const double byte_per_char = data_size / character_count; 
+        const double byte_per_char = data_size / character_count;
 
         const double gbs_avs = data_size / (all.total.elapsed_ns()/all.iterations);
         const double error_margin = (gbs-gbs_avs)/gbs_avs * 100;
