+++ src/avx512bw/avx512bw-undefmacros.inl.cpp
@@ -1,0 +1,2 @@
+#undef TRANSCODE16
+#undef STORE_ASCII
