+++ src/generic/utf8_to_utf16/valid_utf8_to_utf16.h
@@ -60,8 +60,8 @@ simdutf_warn_unused size_t convert_valid(const char* input, size_t size,
         utf8_end_of_code_point_mask >>= consumed;
       }
       // At this point there may remain between 0 and 12 bytes in the
-      // 64-byte block.These bytes will be processed again. So we have an 
-      // 80% efficiency (in the worst case). In practice we expect an 
+      // 64-byte block.These bytes will be processed again. So we have an
+      // 80% efficiency (in the worst case). In practice we expect an
       // 85% to 90% efficiency.
     } else {
       in.store_ascii_as_utf16(utf16_output);
